docs(samples): splits product search samples into separate files
