+++ samples/productSearch/importProductSets.js
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC
+// Copyright 2020 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,70 +13,65 @@
 // limitations under the License.
 
 'use strict';
-// [START vision_product_search_import_product_images]
-async function importProductSets(projectId, location, gcsUri) {
-  // Imports the Google Cloud client library
-  // [START vision_product_search_tutorial_import]
-  const vision = require('@google-cloud/vision');
-  // [END vision_product_search_tutorial_import]
-  // Creates a client
-  const client = new vision.ProductSearchClient();
 
-  /**
-   * TODO(developer): Uncomment the following line before running the sample.
-   */
-  // const projectId = 'Your Google Cloud project Id';
-  // const location = 'A compute region name';
-  // const gcsUri = 'Google Cloud Storage path of the input image'';
+function main(projectId, location, gcsUri) {
+  // [START vision_product_search_import_product_images]
+  async function importProductSets() {
+    // Imports the Google Cloud client library
+    // [START vision_product_search_tutorial_import]
+    const vision = require('@google-cloud/vision');
+    // [END vision_product_search_tutorial_import]
+    // Creates a client
+    const client = new vision.ProductSearchClient();
 
-  // A resource that represents Google Cloud Platform location.
-  const projectLocation = client.locationPath(projectId, location);
+    /**
+     * TODO(developer): Uncomment the following line before running the sample.
+     */
+    // const projectId = 'Your Google Cloud project Id';
+    // const location = 'A compute region name';
+    // const gcsUri = 'Google Cloud Storage path of the input image'';
 
-  // Set the input configuration along with Google Cloud Storage URI
-  const inputConfig = {
-    gcsSource: {
-      csvFileUri: gcsUri,
-    },
-  };
+    // A resource that represents Google Cloud Platform location.
+    const projectLocation = client.locationPath(projectId, location);
 
-  // Import the product sets from the input URI.
-  const [response, operation] = await client.importProductSets({
-    parent: projectLocation,
-    inputConfig: inputConfig,
-  });
+    // Set the input configuration along with Google Cloud Storage URI
+    const inputConfig = {
+      gcsSource: {
+        csvFileUri: gcsUri,
+      },
+    };
 
-  console.log('Processing operation name: ', operation.name);
+    // Import the product sets from the input URI.
+    const [response, operation] = await client.importProductSets({
+      parent: projectLocation,
+      inputConfig: inputConfig,
+    });
 
-  // synchronous check of operation status
-  const [result] = await response.promise();
-  console.log('Processing done.');
-  console.log('Results of the processing:');
+    console.log('Processing operation name: ', operation.name);
 
-  for (const i in result.statuses) {
-    console.log('Status of processing ', i, 'of the csv:', result.statuses[i]);
+    // synchronous check of operation status
+    const [result] = await response.promise();
+    console.log('Processing done.');
+    console.log('Results of the processing:');
 
-    // Check the status of reference image
-    if (result.statuses[i].code === 0) {
-      console.log(result.referenceImages[i]);
-    } else {
-      console.log('No reference image.');
+    for (const i in result.statuses) {
+      console.log(
+        'Status of processing ',
+        i,
+        'of the csv:',
+        result.statuses[i]
+      );
+
+      // Check the status of reference image
+      if (result.statuses[i].code === 0) {
+        console.log(result.referenceImages[i]);
+      } else {
+        console.log('No reference image.');
+      }
     }
+    // [END vision_product_search_import_product_images]
   }
-  // [END vision_product_search_import_product_images]
+  // [END vision_product_search_import_product_set]
+  importProductSets();
 }
-// [END vision_product_search_import_product_set]
-
-require(`yargs`) // eslint-disable-line
-  .demand(1)
-  .command(
-    `importProductSets <projectId> <location> <gcsUri>`,
-    `Import a Product Set`,
-    {},
-    opts => importProductSets(opts.projectId, opts.location, opts.gcsUri)
-  )
-  .example(`node $0 COMMAND ARG`)
-  .wrap(120)
-  .recommendCommands()
-  .epilogue(`For more information, see https://cloud.google.com/vision/docs`)
-  .help()
-  .strict().argv;
+main(...process.argv.slice(2));
