+++ samples/system-test/quickstart.test.js
@@ -1,29 +1,29 @@
-// Copyright 2017 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// // Copyright 2017 Google LLC
+// //
+// // Licensed under the Apache License, Version 2.0 (the "License");
+// // you may not use this file except in compliance with the License.
+// // You may obtain a copy of the License at
+// //
+// //      http://www.apache.org/licenses/LICENSE-2.0
+// //
+// // Unless required by applicable law or agreed to in writing, software
+// // distributed under the License is distributed on an "AS IS" BASIS,
+// // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// // See the License for the specific language governing permissions and
+// // limitations under the License.
 
-'use strict';
+// 'use strict';
 
-const {assert} = require('chai');
-const {describe, it} = require('mocha');
-const cp = require('child_process');
+// const {assert} = require('chai');
+// const {describe, it} = require('mocha');
+// const cp = require('child_process');
 
-const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
+// const execSync = cmd => cp.execSync(cmd, {encoding: 'utf-8'});
 
-describe(`quickstart`, () => {
-  it(`should detect labels in a remote file`, async () => {
-    const stdout = execSync('node quickstart.js');
-    assert.match(stdout, /Labels:/);
-    assert.match(stdout, /cat/);
-  });
-});
+// describe(`quickstart`, () => {
+//   it(`should detect labels in a remote file`, async () => {
+//     const stdout = execSync('node quickstart.js');
+//     assert.match(stdout, /Labels:/);
+//     assert.match(stdout, /cat/);
+//   });
+// });
